Regenerated file
Keep an eye on this. It looks like this file uses comment hooks in the LESS files to generate the Customizer page.  We probably haven't set those hooks up properly, so this will need to be fixed if we decide to provide a customization page.  Luckily it's hidden for now.